Moved classes to seperate files, setup project for pytest and added sample test
