+++ README.md
@@ -95,7 +95,12 @@ TODO: source of data (f_meanEMB, f_sigEMB, EMBmort, env_meanEMB, env_sigEMB)
 ## Usage
 To run the WIP refactored model, enter the ```refactor``` directory and run:
 
-```python SeaLiceMgmt.py folder_name exp_name```
+```python -m src.SeaLiceMgmt folder_name exp_name```
+
+To test, also from ```refactor``` directory, run:
+
+```pytest```
+
 ### Requirements
 Details on required packages and versions can be found in ```environment.yml``` which can be used to create a [conda](https://conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html) environment for the project.
